requirements.txt
tensorflow
transformers
torch
transformers
flask
spacy
scikit-learn
pandas
matplotlib
seaborn
pyyaml
astroid
pytest